Unbundle autocapturing of pageviews
